refactor: убрать лайки из фильмов и друзей из юзеров, уменьшить количество обращений к БД, перенести всю сортировку в БД

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -3,20 +3,18 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
 import ru.yandex.practicum.filmorate.storage.GenreStorage;
-import ru.yandex.practicum.filmorate.storage.LikesStorage;
+import ru.yandex.practicum.filmorate.storage.UserStorage;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
@@ -24,33 +22,25 @@ import java.util.stream.Collectors;
 public class FilmService {
     private final FilmStorage filmStorage;
     private final GenreStorage genreStorage;
-    private final LikesStorage likesStorage;
-    private final UserService userService;
+    private final UserStorage userStorage;
     private static final LocalDate FIRST_FILM_RELEASE_DATE = (LocalDate.of(1895, 12, 28));
 
     @Autowired
-    public FilmService(FilmStorage filmStorage, GenreStorage genreStorage,
-                       LikesStorage likesStorage, UserService userService) {
+    public FilmService(FilmStorage filmStorage, GenreStorage genreStorage, UserStorage userStorage) {
         this.filmStorage = filmStorage;
         this.genreStorage = genreStorage;
-        this.likesStorage = likesStorage;
-        this.userService = userService;
+        this.userStorage = userStorage;
     }
 
     public List<Film> getFilms() {
         Map<Integer, List<Genre>> filmsGenres = genreStorage.getAllFilmsGenres();
-        Map<Integer, Set<Integer>> filmsLikes = likesStorage.getAllFilmsLikes();
 
         return filmStorage.getFilms().stream()
                 .map(film -> {
                     film.setGenres(new ArrayList<>());
-                    film.setLikes(new HashSet<>());
 
                     if (filmsGenres.containsKey(film.getId())) {
                         film.setGenres(filmsGenres.get(film.getId()));
-                    }
-                    if (filmsLikes.containsKey(film.getId())) {
-                        film.setLikes(filmsLikes.get(film.getId()));
                     }
                     return film;
                 })
@@ -58,40 +48,28 @@ public class FilmService {
     }
 
     public List<Film> getMostPopularFilms(int count) {
-        Map<Integer, Set<Integer>> filmsLikes = likesStorage.getMostPopularFilmsLikes(count);
-        List<Integer> filmsIds = new ArrayList<>(filmsLikes.keySet());
-        Map<Integer, List<Genre>> filmsGenres = genreStorage.getGenresByFilmsIds(filmsIds);
+        List<Film> films = filmStorage.getMostPopularFilms(count);
+        Map<Integer, List<Genre>> genres = genreStorage.getGenresByFilmsIds(
+                films.stream().map(Film::getId).collect(Collectors.toList()));
 
-        List<Film> popularFilms = filmStorage.getFilmsByIds(filmsIds);
-
-        if (popularFilms.isEmpty()) {
-            return getFilms().stream()
-                    .limit(count)
-                    .collect(Collectors.toList());
-        }
-
-        return popularFilms.stream()
+        return films.stream()
                 .map(film -> {
                     film.setGenres(new ArrayList<>());
-                    film.setLikes(new HashSet<>());
 
-                    if (filmsGenres.containsKey(film.getId())) {
-                        film.setGenres(filmsGenres.get(film.getId()));
-                    }
-                    if (filmsLikes.containsKey(film.getId())) {
-                        film.setLikes(filmsLikes.get(film.getId()));
+                    if (genres.containsKey(film.getId())) {
+                        film.setGenres(genres.get(film.getId()));
                     }
                     return film;
                 })
-                .sorted(Comparator.comparingInt((Film film) -> film.getLikes().size()).reversed())
                 .collect(Collectors.toList());
     }
 
     public Film getFilmById(int id) {
-        filmStorage.checkFilmId(id);
-        Film film = filmStorage.getFilmById(id);
+        if (!filmStorage.filmExists(id)) {
+            throw new NotFoundException(String.format("Film %d is not found", id));
+        }
 
-        film.setLikes(likesStorage.getLikesByFilmId(id));
+        Film film = filmStorage.getFilmById(id);
         film.setGenres(genreStorage.getGenresByFilmId(id));
 
         return film;
@@ -103,10 +81,8 @@ public class FilmService {
         Film filmReturned = filmStorage.addFilm(film);
 
         genreStorage.updateFilmGenres(film);
-        likesStorage.updateFilmLikes(film);
 
         filmReturned.setGenres(genreStorage.getGenresByFilmId(filmReturned.getId()));
-        filmReturned.setLikes(likesStorage.getLikesByFilmId(filmReturned.getId()));
 
         log.debug("POST request handled: new film added");
         return filmReturned;
@@ -114,39 +90,51 @@ public class FilmService {
 
     public Film updateFilm(Film film) {
         validateReleaseDate(film);
-        filmStorage.checkFilmId(film.getId());
+
+        if (!filmStorage.filmExists(film.getId())) {
+            throw new NotFoundException(String.format("Film %d is not found", film.getId()));
+        }
 
         Film filmReturned = filmStorage.updateFilm(film);
 
         genreStorage.updateFilmGenres(film);
-        likesStorage.updateFilmLikes(film);
 
         filmReturned.setGenres(genreStorage.getGenresByFilmId(filmReturned.getId()));
-        filmReturned.setLikes(likesStorage.getLikesByFilmId(filmReturned.getId()));
 
         log.debug(String.format("PUT request handled: film %d updated", film.getId()));
         return filmReturned;
     }
 
     public void deleteFilm(int id) {
-        filmStorage.checkFilmId(id);
+        if (!filmStorage.filmExists(id)) {
+            throw new NotFoundException(String.format("Film %d is not found", id));
+        }
+
         filmStorage.deleteFilm(id);
         log.debug(String.format("DELETE request handled: film %d deleted", id));
     }
 
     public void addLike(int filmId, int userId) {
-        filmStorage.checkFilmId(filmId);
-        userService.checkUserId(userId);
+        if (!filmStorage.filmExists(filmId)) {
+            throw new NotFoundException(String.format("Film %d is not found", filmId));
+        }
+        if (!userStorage.userExists(userId)) {
+            throw new NotFoundException(String.format("User %d is not found", userId));
+        }
 
-        likesStorage.addLike(filmId, userId);
+        filmStorage.addLike(filmId, userId);
         log.debug(String.format("PUT request handled: like from user %d added to film %d", userId, filmId));
     }
 
     public void deleteLike(int filmId, int userId) {
-        filmStorage.checkFilmId(filmId);
-        userService.checkUserId(userId);
+        if (!filmStorage.filmExists(filmId)) {
+            throw new NotFoundException(String.format("Film %d is not found", filmId));
+        }
+        if (!userStorage.userExists(userId)) {
+            throw new NotFoundException(String.format("User %d is not found", userId));
+        }
 
-        likesStorage.deleteLike(filmId, userId);
+        filmStorage.deleteLike(filmId, userId);
         log.debug(String.format("DELETE request handled: like from user %d deleted from film %d", userId, filmId));
     }
